+++ examples/run_examples.sh
@@ -1,0 +1,13 @@
+#!/bin/bash
+
+FILES="common/*.py torch/*.py"
+
+for file in $FILES
+do
+    echo "--------------------------------------------------------------------------------------"
+    echo "#---# run python $file"
+    python "$file"
+    if [ $? -ne 0 ]; then
+        exit 1
+    fi
+done